+++ include/sqlpp11/sqlite3/prepared_statement.h
@@ -37,7 +37,7 @@
 #include <sqlpp11/exception.h>
 #include <sqlpp11/sqlite3/export.h>
 
-#include <sqlpp11/sqlite3/prepared_statement_handle.h>
+#include <sqlpp11/sqlite3/detail/prepared_statement_handle.h>
 
 #ifdef _MSC_VER
 #pragma warning(push)
